# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "hearsay",
    "description": "Local-first meeting-note transcriber. The core serves this API on `127.0.0.1` only, for the bundled web UI.\n\n**Conventions**\n\n- Base path is `/api`. Every request needs the session token; see the `session_token` scheme.\n- List endpoints return `{ total, page, page_size, items }` and take `page` (at least 1, default 1) and `page_size` (1 to 200, default 50).\n- Errors return `{ \"detail\": \"...\" }`. Database and internal failures render as a literal \"internal error\" — details go to the log only.\n- Times are ISO 8601. Segment `start_s` and `end_s` are seconds relative to the meeting start, not wall clock.\n\n**Not in this document**\n\nThree served routes have no OpenAPI representation and are specified in `docs/api.md`: the live transcript WebSocket at `/ws/meetings/{id}`, meeting audio at `/api/meetings/{id}/audio` (a byte stream with `Range` support), and `/`, which serves the UI with the token injected.",
    "license": {
      "name": "Apache-2.0 WITH Commons-Clause",
      "url": "https://github.com/rwiseman87/hearsay/blob/main/LICENSE"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/meetings/{id}/stop": {
      "post": {
        "tags": [
          "meetings"
        ],
        "summary": "Stop and finalize a meeting.",
        "description": "Stops capture, flushes the pipeline, rewrites `transcript.md` in timestamp order, and stamps\n`ended_at`. Returns the updated meeting immediately: when a refine or notes step will run the\nstatus is `refining`, and it flips to `finalized` once that background work completes.",
        "operationId": "stop_meeting",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The finalized (or refining) meeting",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingRead"
                }
              }
            }
          },
          "404": {
            "description": "No such meeting"
          },
          "503": {
            "description": "Capture engine unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MeetingRead": {
        "type": "object",
        "description": "A meeting row for the API.",
        "required": [
          "id",
          "title",
          "folder",
          "status",
          "started_at",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "ended_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "folder": {
            "type": "string",
            "description": "The on-disk recordings-directory name (not an organizational folder). To move a meeting\nbetween user folders, use `folder_id`."
          },
          "folder_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "The organizational [`FolderRead`] this meeting is filed under, or `null` when unfiled."
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "started_at": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/MeetingStatus"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MeetingStatus": {
        "type": "string",
        "description": "Lifecycle state of a meeting (lowercase on the wire): `recording` while live, `refining` while the\npost-stop refine + transcript write run in the background, then `finalized`.",
        "enum": [
          "recording",
          "refining",
          "finalized"
        ]
      },
...
    },
    "securitySchemes": {
      "session_token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Per-process session token. `make rust-serve` prints it in the `open:` URL; the packaged app passes it to the webview through a handshake file. Channels that cannot set a header — the WebSocket and the audio element — take `?token=` instead."
      }
    }
  },
  "security": [
    {
      "session_token": []
    }
  ],
...
}